Init Shield Workflow
