+++ core/api.go
@@ -27,6 +27,10 @@ const APIVersion = 2
 
 func (core *Core) ServeHTTP(w http.ResponseWriter, req *http.Request) {
 	switch {
+	case match(req, `GET /init.js`):
+		core.initJS(w, req)
+		return
+
 	case match(req, `POST /auth/login`):
 		core.authLogin(w, req)
 		return
@@ -459,6 +463,41 @@ func (core *Core) mustBeUnlocked(w http.ResponseWriter) bool {
 	return true
 }
 
+func (core *Core) initJS(w http.ResponseWriter, req *http.Request) {
+	fmt.Fprintf(w, "// init.js\n")
+	fmt.Fprintf(w, "var $global = {}\n")
+
+	health, err := core.checkHealth()
+	if err != nil {
+		log.Errorf("init.js: failed to check health of SHIELD core: %s", err)
+		fmt.Fprintf(w, "// failed to determine health of SHIELD core...\n")
+	} else {
+		b, err := json.Marshal(health)
+		if err != nil {
+			log.Errorf("init.js: failed to marshal health data into JSON: %s", err)
+			fmt.Fprintf(w, "// failed to determine health of SHIELD core...\n")
+			fmt.Fprintf(w, "$global.hud = {\"health\":{\"shield\":{\"core\":\"down\"}}};\n")
+		} else {
+			fmt.Fprintf(w, "$global.hud = %s;\n", string(b));
+		}
+	}
+
+	id, _ := core.checkAuth(req)
+	if (id == nil) {
+		fmt.Fprintf(w, "$global.auth = {\"unauthenticated\":true};\n")
+	} else {
+		b, err := json.Marshal(id)
+		if err != nil {
+			log.Errorf("init.js: failed to marhsal auth id data into JSON: %s", err)
+			fmt.Fprintf(w, "// failed to determine user authentication state...\n")
+			fmt.Fprintf(w, "$global.auth = {\"unauthenticated\":true};\n")
+		} else {
+			fmt.Fprintf(w, "$global.auth = %s;\n", string(b))
+		}
+	}
+	w.WriteHeader(200)
+}
+
 func (core *Core) authLogin(w http.ResponseWriter, req *http.Request) {
 	// only applies to the local backend configuration
 	// so we can assume that `username` and `password` are set in the
@@ -519,87 +558,15 @@ func (core *Core) authLogout(w http.ResponseWriter, req *http.Request) {
 	w.WriteHeader(302)
 }
 
-type authResponseTenant struct {
-	UUID uuid.UUID `json:"uuid"`
-	Name string    `json:"name"`
-	Role string    `json:"role"`
-}
-type authResponseUser struct {
-	Name    string `json:"name"`
-	Account string `json:"account"`
-	Backend string `json:"backend"`
-	SysRole string `json:"sysrole"`
-}
-type authResponse struct {
-	User    authResponseUser     `json:"user"`
-	Tenants []authResponseTenant `json:"tenants"`
-}
-
 func (core *Core) authID(w http.ResponseWriter, req *http.Request) {
-	cookie, err := req.Cookie(SessionCookieName)
-	if err != nil {
-		if err == http.ErrNoCookie {
-			log.Debugf("no session cookie ('%s') found in request; treating as unauthenticated", SessionCookieName)
-			JSONLiteral(w, `{"unauthenticated":true}`)
-			w.WriteHeader(500)
-			return
-		}
-		log.Debugf("failed to extract session cookie ('%s') from request: %s", SessionCookieName, err)
+	id, _ := core.checkAuth(req)
+	if id == nil {
 		JSONLiteral(w, `{"unauthenticated":true}`)
-		w.WriteHeader(500)
+		w.WriteHeader(200)
 		return
 	}
 
-	log.Debugf("retrieving user account for session '%s'", cookie.Value);
-	user, err := core.DB.GetUserForSession(cookie.Value)
-	if err != nil {
-		log.Debugf("failed to retrieve user account for session '%s': %s", cookie.Value, err)
-		JSONLiteral(w, `{"unauthenticated":true}`)
-		w.WriteHeader(500)
-		return
-	}
-	if user == nil {
-		log.Debugf("failed to retrieve user account for session '%s': database did not throw an error, but returned a nil user", cookie.Value)
-		JSONLiteral(w, `{"unauthenticated":true}`)
-		w.WriteHeader(500)
-		return
-	}
-
-	answer := authResponse{
-		User: authResponseUser{
-			Name:    user.Name,
-			Account: user.Account,
-			Backend: user.Backend,
-			SysRole: user.SysRole,
-		},
-	}
-
-	memberships, err := core.DB.GetMembershipsForUser(user.UUID)
-	if err != nil {
-		log.Debugf("failed to retrieve tenant memberships for user %s@%s (uuid %s): %s",
-			user.Account, user.Backend, user.UUID.String(), err)
-		JSONLiteral(w, `{"unauthenticated":true}`)
-		w.WriteHeader(500)
-		return
-	}
-
-	answer.Tenants = make([]authResponseTenant, len(memberships))
-	for i, membership := range memberships {
-		answer.Tenants[i].UUID = membership.TenantUUID
-		answer.Tenants[i].Name = membership.TenantName
-		answer.Tenants[i].Role = membership.Role
-	}
-
-	if answer.User.Backend == "local" {
-		answer.User.Backend = "SHIELD"
-
-	} else {
-		log.Debugf("looking up auth provider configuration for '%s'", user.Backend)
-		if p, err := core.FindAuthProvider(answer.User.Backend); err == nil {
-			answer.User.Backend = p.DisplayName()
-		}
-	}
-	JSON(w, answer)
+	JSON(w, id)
 }
 
 func (core *Core) v1Ping(w http.ResponseWriter, req *http.Request) {
@@ -1564,107 +1531,12 @@ func (core *Core) v1CancelTask(w http.ResponseWriter, req *http.Request) {
   }
 
 */
-type v2StorageHealth struct {
-	UUID    uuid.UUID `json:"uuid"`
-	Name    string    `json:"name"`
-	Healthy bool      `json:"healthy"`
-}
-type v2JobHealth struct {
-	UUID    uuid.UUID `json:"uuid"`
-	Target  string    `json:"target"`
-	Job     string    `json:"job"`
-	Healthy bool      `json:"healthy"`
-}
-type v2Health struct {
-	SHIELD struct {
-		Version string `json:"version"`
-		IP      string `json:"ip"`
-		FQDN    string `json:"fqdn"`
-		Env     string `json:"env"`
-		Color   string `json:"color"`
-	} `json:"shield"`
-	Health struct {
-		Core    string `json:"core"`
-		Storage bool   `json:"storage_ok"`
-		Jobs    bool   `json:"jobs_ok"`
-	} `json:"health"`
-
-	Storage []v2StorageHealth `json:"storage"`
-	Jobs    []v2JobHealth     `json:"jobs"`
-
-	Stats struct {
-		Jobs     int `json:"jobs"`
-		Systems  int `json:"systems"`
-		Archives int `json:"archives"`
-		Storage  int `json:"storage"`
-		Daily    int `json:"daily"`
-	} `json:"stats"`
-}
-
 func (core *Core) v2GetHealth(w http.ResponseWriter, req *http.Request) {
-	var health v2Health
-
-	health.SHIELD.Version = Version
-	health.SHIELD.Env = os.Getenv("SHIELD_NAME")
-	health.SHIELD.IP = core.ip
-	health.SHIELD.FQDN = core.fqdn
-
-	health.Health.Storage = true
-	stores, err := core.DB.GetAllStores(nil)
+	health, err := core.checkHealth()
 	if err != nil {
 		bail(w, err)
 		return
 	}
-	health.Storage = make([]v2StorageHealth, len(stores))
-	for i, store := range stores {
-		health.Storage[i].UUID = store.UUID
-		health.Storage[i].Name = store.Name
-		health.Storage[i].Healthy = true // FIXME
-		if !health.Storage[i].Healthy {
-			health.Health.Storage = false
-		}
-	}
-
-	health.Health.Jobs = true
-	jobs, err := core.DB.GetAllJobs(nil)
-	if err != nil {
-		bail(w, err)
-		return
-	}
-	health.Jobs = make([]v2JobHealth, len(jobs))
-	for i, job := range jobs {
-		health.Jobs[i].UUID = job.UUID
-		health.Jobs[i].Target = job.TargetName
-		health.Jobs[i].Job = job.Name
-		health.Jobs[i].Healthy = job.Healthy()
-
-		if !health.Jobs[i].Healthy {
-			health.Health.Jobs = false
-		}
-	}
-	health.Stats.Jobs = len(jobs)
-
-	if health.Health.Core, err = core.vault.Status(); err != nil {
-		bail(w, err)
-		return
-	}
-
-	if health.Stats.Systems, err = core.DB.CountTargets(nil); err != nil {
-		bail(w, err)
-		return
-	}
-
-	if health.Stats.Archives, err = core.DB.CountArchives(nil); err != nil {
-		bail(w, err)
-		return
-	}
-
-	if health.Stats.Storage, err = core.DB.ArchiveStorageFootprint(nil); err != nil {
-		bail(w, err)
-		return
-	}
-
-	health.Stats.Daily = 0 // FIXME
 
 	JSON(w, health)
 }
